feat: Split comment component
